backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
qdrant-client==1.11.3
google-generativeai==0.8.3
PyPDF2==3.0.1
torch==2.5.1
sentence-transformers==3.1.1
better-exceptions==0.3.3
pytest==8.3.3
pytest-asyncio==0.23.7
httpx==0.27.2